Style: Shift creation of Task variable to addTask function
Task is created in startChat() function.

Shifting the creation of Task variable from startChat function in
Beefy class to the addTask function in TaskList class, will follow the
proper coding standards.

Let's change the function addTask in TaskList to accept the task
description and create the Task variable in it.