Refactor code: Add command classes to improve code quality
All commands are handled inside Beefy, causing code to be
unreadable as there are multiple levels of abstraction.

Adding multiple command classes help to distribute the code out into
multiple classes and improves readability.

Let's create an interface Command, and create commands for each of
the different command type to implement the interface.