Bug Fix: Fix Storage class readDisk function
Does not create directory to store storage text file when there is no
directory found, causing IO exceptions to occur upon writing to disk.

Creating a file when there is no directory found will prevent IO
exceptions from occurring.

Let's create a function that will be called to create the directory and
file when there is a FileNotFoundException.