Refactor code: Refactor code for adding of deadline/event
An empty task description can be added into the task list.

Making the code check if the task to be added has no description
will be able to ensure clean data storage.

Let's make the DeadlineCommand and EventCommand constructors throw
a BeefyException when an empty task description is passed into it.