fix: Fix incorrect printing of total tasks
the `totalTasksCreated` was used previously as the delete functionality
was not introduced, and so the total number of tasks was the number
of the last item on the list. When delete was introduced, it was
decided that all new tasks created will have a unique Task ID. Hence,
there is now a need to separate check the number of active tasks in the
list.